Change default encoding to binary, instead of hex
This patch makes a few changes, but most importantly, we define a breaking change to make the default encoding of an RLP structure the binary encoding (e.g. <<0x55, 0x66, 0x77>>) instead of hex (e.g. "556677"). This is because the binary encoding is faster and more appropriate when encoding and decoding RLP from the blockchain. We allow a user of ExRLP to return to hex encoding by specifying `encoding: :hex` in the options hash.

Additionally, we merge the test cases from another RLP library and add typespecs for dialyzer.