Upgrade Merkle Patricia Trie Implementation
This patch upgrades the merkle patricia trie to use the code adopted from another implementation (previously part of `hayesgm/exthereum`'s umbrella project). The primary purpose of this change is to make sure we have a trie that is fully compatible with the rest of the libraries for Exthereum. We have chosen this fork as it provides a stronger set of test cases and supports both :ets and :leveldb (the standard for the Ethereum community). This presents a breaking change for all clients to this library, and as such, we have updated the version to 0.2.0.