Update to add typspecs and more unit tests
This patch (while it looks large) is a relatively minor set of changes to make sure that each function is properly typespeced and has additional unit tests. We add documentation for each public function and module, and generally tried to make the intent of the code clearer if possible. Other than that, we change very little actual behavior of the code in this commit.

Note, this is a WIP as test cases are not currently passing.