[WIP] Start to add support for chains
Chains are a feature of Ethereum which group a genesis block and a set of parameters specific to a chain. For instance, testchains such as Ropsten have a specific set of accounts with starting balance, as well as the inclusion of EIP (forks) at different block numbers (e.g. Homestead starts at block 10). This is essential that we load this information and pass it to different functions if we want to support and match data we receive (be it from a test chain or Homestead of Serenity). We will need to pass chain information to a number of functions (including to the EVM as we will toggle flags such as changes in gas price based on the chain we're in).