Add a test case that loads live Ropsten chain data
This patch adds a test case that loads live Ropsten chain data. This can be used to easily test our progress as it's our most definitive way of knowing whether or not our client functions. Right now, we can process the first 11 blocks, but in the 12th block, we start to run transactions.