Add in a full debugger to the EVM
This patch adds a first-class debugger to the EVM. As we are starting to run real transactions from the blockchain, it's becoming more and more important that we can walk through the execution of these contracts to understand the effects as they are run. I expect this debugger to grow into a full first-class debugger (imagine being able to change code inline as you debug it, move the PC, set more interesting breakpoints, etc). This debugger is currently working and lets you set and walk through breakpoints on live code.