Add Stream Cipher and other minor improvements
This patch adds the Stream Cipher for AES, which is used in RLPx's handshake protocol. We also opt to store public keys in a raw format, instead of starting with a 0x04 byte indicating an octet_string in DER format. This distinction is always confusing, but messed up our encoding and packet size.