[WIP] Start to add RLPx
This patch is very much a work in progress, but I wanted to give time to start to review early on. The idea here is to implement the EIP-8 encoding of the handshake messages now that ExCrypto is coming along with ECIES. Once we have a handshake algorithm (and ack), we should be able to start to move on to protocol negotiation.