Add Eth Wire Syncing Module
This patch adds the ability to start syncing from remote nodes. We basically start one "sync" process that's responsible for trying to interpret messages from remote peers to form a coherent chain. We spawn any number of (okay, for now, one) TCP connections to remote nodes so that we can ask for headers and bodies for blocks. As the sync module gets BlockHeaders (and BlockBodies), it checks if it has a sufficient number of confirmations from peers (again, currently one) and if so, it tries to process that block and add it to our local chain. If that works, it asks for more headers / bodies to continue the process. Other than the fact our canonical chain ("Blocktree") isn't stored locally, this should basically outline the entire sync process (short of Fast / Warp).