Upgrade Interfaces to be Compatible with Blockchain
This patch upgrades the interfaces of EVM to make sure that they are fully compatible with an implementation in the Blockchain app. The majority of the changes are minor changes to be compliant with this new version of the API interface. We also eradicate ContractInterface and store everything in AccountInterface.