Add define_common_tests macro
We add a `define_common_tests/2` macro to handle the new directory
structure of ethereum common tests. For now, we have followed Mason's
implementation very closely. But there are a few things that we would
like to change,

* We should try to avoid having `try rescue` parts in `parse_test`
function in `Blockchain.TransactionTest`. That seems like a code smell
-- we are rescuing `ArgumentError -> nil`.

* Mason's implementation also has a `if parsed_tests.transaction.v < 30`
check with a comment to handle the case where `v` is > 30. We should
handle that before we merge.

* Last, the old code ignored invalid tests. It had code like the
following,

```
if src_data["expect"] == "invalid" do
  # TODO: Include checks of "invalid" tests
  Logger.debug("Skipping invalid transaction test: TransactionTests/EIP555 - #{test_subset} - #{test_name}")
```

We should either handle those cases, or we're going to have to do
something similar (though I'm not sure if we're even getting that data
anymore.

Last note - we are leaving the block_test using the `eth_test` macro for
now. We should update that to use the new macro before this is done.