Fix block_tests
Two tests were failing in block_test.exs and one doc spec in Block.ex.
The doc spec was failing as were other doc tests. It seems that the
`log` was changed to return a list instead of a string. So a simple
replacement of `[]` for `""` fixes those tests.

The other two tests were failing due to a misspelled `mixHash`. It had a
lower case `h`, making it `mixhash` and causing it to be `nil` (since it
was getting it from a map.