Add chain_id for forks that need it
According to the [EIP155 specification](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-155.md),
a chain_id is used when signing the transaction. So any chains after the
EIP155 fork need to include that chain id when signing the transaction.

The implementation code already handles that. But the tests were
ignoring any common tests that had a `v` value greater than `30`. We fix
that here by passing the `mainnet`'s `chain_id` of `1` for any chains
that came after EIP155.

Because we now handle that, the logic for all common tests generated
from the `TransactionTests` directory can be combined into a single use
of `define_common_tests/3` (note we continue to ignore the tests we were
already excluding)