Fix formatting on code comments related to equations
Fix formatting on code comments that are describing the elements of a
structure relative to an equation found in the yellow paper.

The automatic formatting puts any inline comment on a line above the
code. When those comments are inline with arguments to a function, it
has the strange effect of making it harder to read the function
signature and what the comment is referencing.

We update three instances where this was the case to make it a single
comment above the function signature that explains what all of the
arguments are.