Improve code formatting of holistic validity check
The automatic formatting for the holistic validity check in
`lib/blockchain/block.ex` was extremely nested since it was doing a lot
of `if else` statements and concatenating lists. We update this to
extract each validity check to its own private function, and we collect
the errors as we go along.

One change made here (as will be noted by the change in the doc test) is
that instead of concatenating the errors like so,

```elixir
[:error1] ++ [:error2]
```

we are adding prepending the latest error (since it's technically
faster),

```elixir
[:error2 | [:error1]]
```

which has the adverse effect of returning the errors in opposite order
(hence the change in the order of the results in the doc tests). If the
order of the errors is desired, we can certainly return the errors in
that order by reversing the list or doing the checks in reverse order.