Extract DEVp2p Wire Protocol from ExWire.Adapter.TCP
Closes #72

Why do this?
==============

The `TCP` module handles too many concerns at the moment: it is a
GenServer managing the TCP connection, it handles the aggregation of TCP
data into wire protocol "packets" (e.g. GetBlockHeaders packet), it
performs the crypto handshake, as well as initializing the session of
the Devp2p Wire Protocol (which includes the protocol handshake by
sending and receiving an initial Hello message).

It would be better if the GenServer could simply delegate most
responsibilities to other modules. In this way, we can separate a
runtime concern (i.e. what code should run in which process) from the
logic of our application.

To that end, we extract the logic related to DEVp2p Wire Protocol. We do
not yet extract the logic related to RLPx, which includes the crypto
handshake, as well as the framing of all messages after the handshakes
have been successful. But this may be a good next step.

What was done?
=============

We create a `ExWire.DEVp2p` module to handle all logic associated with
the DEVp2p Wire Procotol. The handshake (`Hello` message) and the notion
of an active session previously lived in the TCP module.

We do not yet create a GenServer to manage the actual handshake with the
peer. For now, all we need is a module that can handle all of the logic.
By passing the state in a `DEVp2p.Session` struct, we can test and
ensure that we can handle a complete handshake without having to keep
the state in a GenServer. In this way, we separate the runtime concerns
(e.g. which process to keep the state in) from the logic of the wire
protocol.

While extracing the `DEVp2p` responsibilities we also refactored some
of the `TCP` code:

- The logic for the TCP Server is complex enough that it makes sense to
move it into its own module. We move it into `TCP.Server` and update
`TCP` (which now the client interface) to call the `Server`.

- We remove the responsibility for handling packets out of TCP. Once the
GenServer has handled the crypto handshake (sending auth and receiving
ack), it should simply forward any packets to subscribers. It should not
handle the packets themselves.