Do not pass unneeded peer's uri host as argument
There were several functions that were taking the peer's host as an
argument solely so they could print it in a `Logger` call. This seems
unnecessary: we should only pass the things that are needed to the
functions. It was also confusing since (a) it made it seem like
functions needed to know about the peer when they didn't, and (b) it
adds a lot of misleading noise when running tests -- logging that
something is being sent over the network when in fact only a struct is
being built.