Fix Secrets.derive_secrets/7 function
The `Secrets.derive_secrets/7` function was throwing an `ArgumentError`
when trying to generate a shared secret. It was untested and for all
intents an purposes, it was unused. So the bug was never caught.

The function failed when trying to generate an ephemeral shared secret.
In order to generate that shared secret, we need our node's ephemeral
private key and the remote node's ephemeral public key.

The function was trying to pass a "raw" ephemeral public key instead.
And when looked at closely, it wasn't even passing the correct "raw"
key. Instead it was using a function that transformed the
`remote_ephemeral_public_key` from raw into `der` (i.e.
`ExthCrypto.Key.raw_to_der()`).

We also add a few tests to ensure that this code is working. Though the
tests are of some value (they caught the error), they are not perfect
since they are simply testing that the values returned are binaries or
references. Better tests would involve using the secrets to do something
that ensures the secrets are indeed correct. Unfortunately, that's more
than I can decipher right now. So for now, I think having some tests is
better than no tests.