Change AuthMsgV4 and AckRespV4 keys
The keys in `AuthMsgV4` and `AckRespV4` are all defined as `remote_*`.
But they can be more accurately named `initiator_*` and `recipient_*`.
Having them named `remote_*` leaves ambiguity since which peer is the
remote peer is dependent on which node we are currently talking about.
That creates confusion.

The most concrete example of this confusion is
`auth_msg.remote_ephemeral_public_key`. When the initiator sends an auth
message, that key will be `nil`. When the recipient receives the
message, they will fill in the `remote_ephemeral_public_key` with the
initiator's public key because the remote is the initiator from their
perspective.

This naming also more closely matches the documentation found in
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-8.md where the
terminology is `initiator-pubk`, `initiator-nonce`,
`recipient-ephermal-pubk`, etc.