Use elixir 1.6.4 in CircleCI
Our builds started failing across several branches. Looking at the
builds, I noticed that the sha for the elixir docker image
(circleci/elixir:1.6) had changed when the builds started working.

Part of the error messages was,

```
mandatory chunk of type 'Atom' not found
```

A quick search revealed that the error is commonly found when compiling
code in a newer version of Erlang/OTP and running it in an older version
(e.g. compiling it in Erlang/OTP 20 and running it in Erlang/OTP 19).

By ssh'ing into CircleCI, I was able to see that we were running Elixir
1.6.5 compiled in Erlang/OTP 19 and we were running in Erlang/OTP 19.

Since we're running in the same version of Erlang as we compiled, I
don't know why we're seeing the error. The compatibility elixir page
suggests that 1.6.5 should work with Erlang/OTP 19.
https://github.com/elixir-lang/elixir/blob/master/lib/elixir/pages/Compatibility%20and%20Deprecations.md
It's possible that since Elixir 1.6.5 supports Erlang/OTP 21, there may
be issues with Erlang 19, contrary to what the compatibility seems to
suggest.

In any case, by setting the CircleCI image to use elixir 1.6.4 (instead
of the latest 1.6.x), we seem to have fixed the problem for now.

In the future, we should use an image that has Erlang/OTP 20 or 21.