Fix Signature recovery in crypto handshake (#148)
During our work for https://github.com/poanetwork/mana/issues/103, we
discovered some tests that fail intermittently. This fixes those issues.

Summary
========

`ExthCrypto.Signature.sign_digest/2` returns a `signature` and a
`recovery_id`. We then concatenate those two when creating the
`AuthMsgV4` struct. When the recipient tries to recover the public key,
however, we are hardcoding the `recovery_id` to `0`. This led to
intermittent failures in our tests (when the recover_id != 0).

To properly handle this, we create two functions in
`ExthCrypto.Signature` to create a compact signature (sig + recover_id)
and to split that compact signature when recovering it.

Using those functions, we fix the bug by no longer hardcoding the `0`
but retrieving the correct recovery id from the auth message, and using
it to recover the public key.

We also try to improve the typespecs in `Signature` since we have
learned that a ECDSA compact signature should always have 65 bytes (64 +
recovery_id).