Fix inbound handshake & add naive TCP inbound implementation
Summary
=======

This commit handles two things that are interrelated. It adds a naive
implementation for an inbound TCP connection and fixes the function that
handles the main crypto handshake when it is initiated by the remote
peer.

Fix inbound handshake
=====================

The main function that was called when a remote node initiated the
handshake (sending an auth message) was `Handshake.try_handle_auth/4`.

`Handshake.try_handle_auth/4`, however, was broken. It had an unrealistic
function signatures (assuming we would have peer data that we can only
obtain after decrypting auth message). It also called
`Handshake.build_ack_resp/2` incorrectly, passing incorrect argument
types and values. This function was untested and practically unused, so
the errors were not apparent at first.

We fix that function and we update the calling code (in TCP.Server) to
use the updated veresion.

Naive implementation of incoming TCP
===================================

We also add a naive implementation for an incoming TCP connection. We
say naive because in this implementation only one client can connect
with our server. The full implementation needs to separate the creation
of the `listen_socket` from the socket created when we accept messages.
We will likely have to span a new process for each connection we
receive.

We also update the `TCP.Server` code since most functions assume the
presence of a peer. At this point, for our naive implementation of the
incoming TCP connection, we do not have the notion of a peer when the
connection is incoming. With a full implementation, we will have the
public key (node id) of the remote peer (obtained from their Auth
message), but we will still not have a `peer` in the sense that is
defined in `ExWire.Struct.Peer` and which is used in outgoing
TCP connections.

Clean up
========

In addition to the above, we also do some clean-up. The main one worth
mentioning is the movement of handshake files.

The `handshake.ex` file was under `lib/ex_wire/handshake/handshake.ex`,
but the module simply defined `ExWire.Handshake`. The more common
pattern is to put `ExWire.Handshake` directly on
`lib/ex_wire/handshake.ex`, so we do that here.