Update TCP.Server handle_info for closed tcp connection
Update TCP.Server.handle_info function when a tcp socket closes. We only
supported outgoing connections (since we assumed the presence of a
`peer` in the state), but we need to handle a closed connection when the
connection was incoming.

This leaves us with branching logic for the message. Most of this
branching logic will be cleaned up when we properly handle all incoming
messages across protocols (RLPx, DEVp2p, and Eth wire protocols).