Handle multiple inbound tcp connections (#165)
Previously we were only accepting one inbound tcp connection. This
commit updates the previous code so that we can handle multiple
connections.

What was done?
==============

Since inbound connections are pretty self-contained, we create a
supervision tree (under the main ex wire supervisor) that is separate
from other trees. There is no reason why a failure on this supervision
tree should crash, say, the node discovery supervision tree.

The top-level supervisor for the tcp inbound communication is the
`TCPListeningSupervisor`. This supervisor supervises an `TCP.Listener`
and a `TCP.InboundConnectionsSupervisor`.

The `TCP.Listener` is a process responsible for opening the tcp
connection on a given port. Once a connection is accepted, it will hand
over control of the connection to another process. Those processes are
started and supervised by the `TCP.InboundConnectionSupervisor`. In this
way, we can have a single process for every inbound connection that is
established.

-----------------------

We also TCP.Server handle_info for closed tcp connection

Update TCP.Server.handle_info function when a tcp socket closes. We only
supported outgoing connections (since we assumed the presence of a
`peer` in the state), but we need to handle a closed connection when the
connection was incoming.

This leaves us with branching logic for the message. Most of this
branching logic will be cleaned up when we properly handle all incoming
messages across protocols (RLPx, DEVp2p, and Eth wire protocols).