Initialize DEVp2p session
Adds simple initialization of DEVp2p session and stores it in the state.

We also start adding a DEVp2p integration test that tests the
communication between two nodes. We test that the initiator is correctly
receiving an ack response (which implies it is sending an auth message
and that the recipient is sending the ack response). And we are testing
that the recipient has sent a Hello message (the DEVp2p protocol
handshake)

A few highlights:

We're asserting these things by tracing the receiving messages in the
initiator process.

Some of the messages we will need to look at include:

Incoming tcp messages,
```
{:tcp, _socket, data}
```

And sending out of packets via GenServer.cast,
```
{:"gen_cast", {:send, packet_data}}
```