Improve Handshake.handle_ack (#169)
While we were doing work for inbound messaging, we improved
`Handshake.handle_auth`, but we left `Handshake.try_handle_ack` so as to
not introduce too much noise into the commits.

Here we do some clean up of that function (renaming it, and improving
the function signature), and we clean up the TCP.Server module which
calls those functions. We also extract `:gen_tcp` functions and reduce
the number of messages the server request of itself.