Add peer struct to inbound TCP.Server
We retrieve peer data and construct a `%Peer{}` struct for inbound TCP
connections. For outgoing connections, this is much easier because we
know which peer we're connecting to. For the inbound connection, it was
a little more difficult.

For background information, the enode url looks like this:

```
enode://hex_node_id@ip_address:port
```

We retrieve the host (ip address) and port peer data from the socket
connection using `:inet.peername/1`, and we retrieve the peer's public
id from decoding the auth message the peer first sends.

Since we can now do that, we can construct a `Peer` struct for all
inbound TCP connections, and we normalize the Logger messages and
function signatures.

The only place where we continue checking whether or not we have a peer
in our state is when the tcp connection is disconnected. It is possible
that a peer will disconnect before we finish the handshake, and thus we
would not have a peer public id at that moment.

As part of this work, we discovered some incorrect typespecs as it
relates to DER encoded public keys. We fix those.

-----------

For more information on enode url format, look at
https://github.com/ethereum/wiki/wiki/enode-url-format