Fix "failed to match header ingress mac"
What was wrong?
===============

The derivation of secrets was incorrect. In particular, we were passing
the incorrect nonce when deriving the ingress/egress macs. We have
separated the `derive_secrets` function into two functions to make it
more clear.

How did we discover it?
========================

In order to correctly fix the error mentioned above, we wanted to improve the
tests for ExWire.Framing.Secrets. The current tests were not very good
as they were only testing the correct data structures were returned but
not the values.

The EIP-8 documentation (found in
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-8.md) has a
section under RLPx Handshake that has some test vectors with which we
can test the validity of the secrets derived. So we use that data (as do
geth and parity) to ensure we are deriving the secrets correctly.

In adding these tests, it was discovered that our implementation was
incorrect.

Refactoring notes:
=================

Since we were already working with these files, we decided to refactor
`Secrets.derive_secrets` to take in a `Handshake` struct. This cleans up
the api noticeably (since all the parameters needed come from the
handshake, and it keeps in line with what geth and parity do as well.

We also update (and refactor) the code in `Handshake` module that makes
use of `Secrets.derive_secrets/1`. We update the code so that all
functions use the `Handshake` struct.