Standardize ExWire.Handshake functions (#235)
ExWire.Handshake has been changing over time as we continuously refactor
while working on the networking layer. This commit updates the functions
to more consistently take an `ExWire.Handshake` struct as a first
argument. It also creates a more even flow whether the node is
initiating the handshake or is the recipient of the handshake. This
mirroring of the code paths makes it easier to reason about and debug.

Why these changes?
===================

In an effort to make the `DEVp2p` session handling easier to reason
about (and hence add), I am trying to tackle the simplification of the
TCP module first.

Eventually, we should have `ExWire.Adapter.TCP` be an "adapter", as the
name suggests. Currently, it contains all the logic that coordinates the
encrypted handshake (and the DEVp2p session initation). Ideally it
should not contain all that logic, and it should simply focus on the tcp
connection, looking more like the `ExWire.Adapter.UDP` module does.