Set DEVp2p session as active (#252)
What?
=====

We update TCP.Server to set DEVp2p session as active, and to check the
session's status when other packets come in. When any other packet comes
in we forward it to the subscribers.

Subscribers is a concept that has been in `TCP.Server` for quite some
time even though we don't really have subscribers (other than the `Sync`
module that is not currently being used). For that reason, I think we
can handle the rest of the messages (i.e. messages other than encrypted
handshake and DEVp2p Hello messages) in one of two ways:

1. We continue having modules/servers that subscribe to TCP connection.
The TCP.Server handles encrypted handshake and protocol handshake, then
forwards all other messages.

2. We simply have the TCP module (or some other module that uses a TCP
connection, which requires a lot of refactoring) handle all messages. It
first ensures that encrypted handshake and protocol handshakes are
performed and then makes use of `handle/1` in packets to deal with data
(essentially removing the `notify_subscribers` part).

It's worth noting that (2) works well for inbound connections (since
they mostly receive a message and respond to it without much regard to
state). But outbound connections (such as the Sync) module we may want
to treat differently.