Clear balance when calling SELFDESTRUCT
What?
=====

After a `SELFDESTRUCT` CPP Ethereum increased the specified refund
address by the current balance in the contract that is marked for
destruction and then set that contract's balance to zero. This lead to a
special case where if the refund address is the same as the contract
address it would cause the ether to essentially vanish from existence,
as described in Quirk 2 of
http://swende.se/blog/Ethereum_quirks_and_vulns.html.

We also add a note that we copied and modified from the CPP
implementation that speaks to the reason why this is done. It seems
there was a consensus issue in the past and it is now tested via the
suicide test we now pass.

----

Clean up Notes:

We remove destroy_account functions. Several modules had
`destroy_account` functions. These are no longer used.

Originally, it seems like we called `ExecEnv.destroy_account` which
would trigger calls to the rest of the `destroy_account` functions. But
in commit c68dcd73dd82401d97654a1374a3b844c6ed0e45, we stopped
destroying the account in the middle of the EVM execution.