Upgrade Rox
Rox was throwing the following error:

````
Unsupported Erlang version.

Is the erlang_nif-sys version up to date in the Cargo.toml?
````

`erlang_nif-sys` was not up to date. We updated it in Rox and open [a PR](https://github.com/urbint/rox/pull/16)

Until our PR gets merged we’ll depend on our fork.

Other changes:

* Updated the circle configuration

Circle was failing to build the newest version of Rox on the [dialyzer step](https://circleci.com/gh/poanetwork/mana/1095)

By restoring from the cache we can use the version of Rox built by the
`build` step which works as expected.