Return hash of block in BLOCKHASH
Our `BLOCKHASH` operation was incorrectly returning the `mix_hash` of
the block instead of the hash. We were also passing an incorrect number
to the `BlockInterface.get_ancestor_header` function, requesting the
block number we wanted instead of the difference. The function, as it is
written, retrieves the ancestor n steps behind.