Cache dialyzer in CI
Dialyzer's PLT is not caching correctly in CI. We make some changes to
the config to fix that. Note that we unpack and pack the PLT into a
directory because it turns out there is a bug in circle ci when trying
to cache the same directory (in this case we would want to cache
`_build` twice). So for the PLT we cache a plt directory we create.

We also want to cache the PLT in relation to our elixir/erlang versions.
For that reason, we add a `.tool-versions` file that keeps track of the
versions we are using.