Reset account on SELFDESTRUCT if it is beneficiary (#311)
It turns out that Frontier and Homestead need to reset an account that
has been marked for destruction (instead of deleting it). They only do
this, however, if the account is the beneficiary account (ie. the
coinbase account).

Resetting the account means that we set the balance to 0 and remove the
code from it.

A commit in Ethereum common tests led us to understand this. I include
it here for future reference:

https://github.com/ethereum/tests/commit/013aa117669dfe82b4ea289a7191fa6658da1334