Do no transfer value if current account is recipient
What?
=====

For an opcode `DELEGATECALL` we should not attempt to transfer the value
(funds) from the sender's account to the recipient since the sender and
value are persisted. In practice, this should be true of all `CALL`,
`CALLCODE`, and `DELEGATECALL` when the current account and the
recipient are the same (I think this is partially true because the
MessageCall in the EVM assumes the current account is the sender).

To this end, we add a check to make sure the address in the current
execution environment is not the same as the recipient. This is also
mentioned in the Yellow Paper (Section 8 - Message Call),

> We define σ1, the first transitional state as the original state but
with the value transferred from sender to recipient, unless s = r.

Other changes:
=============

- Update README to reflect 100% Homestead blockchain tests passing.

- We also add a `EVM.TestFactory` module to help with test setup. The
main interaction with the factory comes through `build/2` which will
then call the factory defined in `factory/2`.