Add Ethereum common difficulty tests
We add Ethereum common tests that test the difficulty calculation
described under the "Block Header Validity" section of the Yellow Paper.

In the common tests, under the `BasicTests` directory, there are several
files that test difficulty. Some are generated manually while others are
refilled automatically. We make use of the file
`difficultyFrontier.json`. It is worth noting that those tests do not
contain difficulty tests only for the Frontier chain. They also test
other forks (like Homestead). I am not sure why they have `Frontier` in
the name since other forks are included.

In order to limit which tests run, so that we can pass tests
progressively as we implement various forks, we check the block number
in the test to ascertain which fork we're testing. For now, we are only
testing Frontier block headers.

Other changes:

- We also move two doctests that were better set as ExUnit tests