Fix difficulty calculation for Frontier fork
By adding the tests difficulty calculation common tests, we discovered
that we our difficulty calculations for Frontier were wrong. The errors
were all an off-by-one type of errors, so we make use of Elixir's
built-in integer division `div/2` function instead of the
`MathHelper.floor` function when calculating a block header's
difficulty.