Add Homestead difficulty common tests (#351)
There is something strange with the common test files for difficulty:

Under the `BasicTests` directory of ethereum common tests, there are
files named `difficultyFrontier.json` and `difficultyHomestead.json`.
Each of them contains tests for Frontier and for Homestead (and other
forks prior to Byzantium as well). The strange thing is that the
Homestead tests in the `difficultyFrontier.json` file fail when tested
against the Homestead implementation, and the same is true of the
Frontier tests in `difficultyHomestead.json` file when tested against
the Frontier implementation.

I don't know why that happens, but my guess is that this is a side
effect of auto-generating the common test files. They probably include
all the tests they have, but they generate them based on a particular
implementation of the difficulty equation. So the `Homestead` blocks in
the `difficultyFrontier.json` file are incorrect because they use the
frontier implementation of the difficulty equation. And the same is true
of the `Frontier` blocks found in the `difficultyHomestead.json` file.

So what we do is simply ignore the Homestead blocks when we are testing
`difficultyFrontier.json` file. And we ignore the Frontier blocks when
we are testing the `difficultyHomestead.json` file.