Prevent collisions on account creation
There was an issue with contract creation that could overwrite
contracts. The solution is to check for a positive nonce or nonempty
code hash in the account that is about to be created. If that's the
case, the contract creation needs to fail instead of overwriting the
account.

For more information see https://github.com/ethereum/EIPs/issues/684