Increment nonce of touched accounts on CREATE
Increment the nonce of an account when it is touched in CREATE
(operation or transaction call). We change the implementation of
`start_nonce` introduced earlier in pull request #353 to
`increment_nonce_on_create` since this seems to more accurately describe
what EIP 161 was referring to in

> a. Account creation transactions and the CREATE operation SHALL, prior
to the execution of the initialisation code, increment the nonce over
and above its normal starting value by one (for normal networks, this
will be simply 1, however test-nets with non-zero default starting
nonces will be different).

Note that the test-nests comment above would be configured at the chain
level not at the EVM fork level.