Fix test_group from path retrieval in state tests (#367)
The test_group variable, which is used to determine known failures, 
was incorrectly returning the wrong part of the path (likely due to us 
changing from a relative path to an absolute path at some point). It 
was usually returning mana. We reverse the index to grab the elements 
from the end of the list so that we always get the group name, regardless 
of where the path starts.