Run Blockchain Tests in different processes (#374)
What changed?
============

We are currently running all blockchain tests (that is the 
ethereum common tests under /BlockchainTests directory) 
in a single ExUnit test. So we can't get any of the automatic 
parallelization that ExUnit's `async: true` provides.

Here we split each hardfork we are testing into a separate 
process. We then combine the results and assert whether 
or not the ExUnit test succeeded or failed as a whole.

In a very limited sample (two CI builds), this seems to shave 
between 6 and 7 minutes off the build.