Run state tests in different processes (#376)
What changed?
=============

We are currently running all state tests (that is the ethereum common 
tests under `/GeneralStateTests` directory) in a single ExUnit test. So 
we can't get any of the automatic parallelization that ExUnit's 
`async: true` provides.

Here we split each hardfork we are testing into a separate process. We 
then combine the results and assert whether or not the ExUnit test 
succeeded or failed as a whole.

We have already done this in `blockchain_tests` so we extract common 
functionality related to spawning processes for each fork into a shared 
module called `AsyncCommonTests`.

From a very brief look at the builds, it looks like this shaves another 4-6 
minutes off of the build.