Use Task.async_stream/3 instead of AsyncCommonTests
After giving this a second look, I realized that the functionality
desired was pretty much what the `Task` module was meant to do. A quick
glance at the docs revealed `Task.async_stream/3` which executes an
anonymous functions in an asynchronous fashion. It's exactly what we
need and so we replace our implementation in `AsyncCommonTests` with
the `Task` implementation.