Split tests by groups instead of forks
In order to improve the build time, we split our tests by forks in
previous commits. When we started using `Task.async_stream`, it became
apparent that we could just split tests by groups of tests rather than
by forks. What we end up doing here is splitting all tests into ten
groups for each fork. We then run those tests asynchronously.

We tested splitting tests into 20 groups but we did not see an
improvement in build times. I think that is due to the fact that the
build can only be as fast as the slowest test we have. And I think one
(or more) of the tests we have take several minutes.