Add Byzantium difficulty calculation (#392)
This relates to EIP 100

What changed?
=============

We add the Byzantium difficulty calculation to `Block.Header` which takes
into account the presence of an ommers hash, and it introduces a fake
block number to delay the ice age. (this is part of EIP-649 

We also update `Blockchain.Block` to calculate the byzantium difficulty
if the block number is after the byzantium transition.

Test changes:

We add Byzantium blokchain tests. We also update the
`generate_blockchain_tests.ex` script to write the failing tests into a
file, which is then read by `blockchain_test.exs`. In this way, we do
not have to copy/paste the failing tests as we go along. We also provide
a separation between the list of failing tests and the code that runs it
(namely the `blockchain_test.exs` file).

Co-authored-by: Matthew Sumner <matt.m.sumner@gmail.com>