Calculate rewards for EIP-649
What changed?
=============

We update the miner and ommer rewards according to specification in
EIP-649.

The updating of the difficulty explained in EIP-649 was already commited
in a previous commit.

Test changes:
============

These changes have passed over 3000 Byzantium tests. So we update the
README with all the latest numbers. We are currently passing 77% of
blockchain tests.

Co-authored-by: Matthew Sumner <matt.m.sumner@gmail.com>